Upgrade monitor model columns in chunks
