+++ tests/ClassUsesTraitTest.php
@@ -5,6 +5,7 @@ namespace romanzipp\QueueMonitor\Tests;
 use romanzipp\QueueMonitor\Services\ClassUses;
 use romanzipp\QueueMonitor\Tests\Support\MonitoredExtendingJob;
 use romanzipp\QueueMonitor\Tests\Support\MonitoredJob;
+use romanzipp\QueueMonitor\Tests\TestCases\TestCase;
 use romanzipp\QueueMonitor\Traits\IsMonitored;
 
 class ClassUsesTraitTest extends TestCase
