+++ tests/MonitorTimeCalculationTest.php
@@ -5,6 +5,7 @@ namespace romanzipp\QueueMonitor\Tests;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
 use romanzipp\QueueMonitor\Models\Monitor;
+use romanzipp\QueueMonitor\Tests\TestCases\TestCase;
 
 class MonitorTimeCalculationTest extends TestCase
 {
